# docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
  CreateIssueRequest:
    properties:
      description:
        type: string
      dueDate:
        type: string
      priority:
        $ref: '#/definitions/IssuePriority'
      status:
        $ref: '#/definitions/IssueStatus'
      title:
        type: string
    type: object
  CreateIssueResponse:
    properties:
      data:
        $ref: '#/definitions/Issue'
    type: object
  CreateProjectRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  CreateProjectResponse:
    properties:
      data:
        $ref: '#/definitions/Project'
    type: object
  GetIssueResponse:
    properties:
      data:
        $ref: '#/definitions/Issue'
    type: object
  GetProjectIssuesListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/ListProjectIssue'
        type: array
      paginator:
        $ref: '#/definitions/Paginator'
    type: object
  GetProjectResponse:
    properties:
      data:
        $ref: '#/definitions/Project'
    type: object
  GetProjectsListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/ListProject'
        type: array
      paginator:
        $ref: '#/definitions/Paginator'
    type: object
  Issue:
    properties:
      createdAt:
        type: string
      description:
        type: string
      dueDate:
        type: string
      id:
        type: integer
      priority:
        $ref: '#/definitions/IssuePriority'
      project:
        $ref: '#/definitions/ProjectShort'
      status:
        $ref: '#/definitions/IssueStatus'
      title:
        type: string
      updatedAt:
        type: string
    type: object
  IssuePriority:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - IssueLowPriority
    - IssueMediumPriority
    - IssueHighPriority
  IssueStatus:
    enum:
    - open
    - in_progress
    - done
    type: string
    x-enum-varnames:
    - IssueOpenStatus
    - IssueInProgressStatus
    - IssueDoneStatus
  ListProject:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  ListProjectIssue:
    properties:
      id:
        type: integer
      priority:
        $ref: '#/definitions/IssuePriority'
      status:
        $ref: '#/definitions/IssueStatus'
      title:
        type: string
    type: object
  Paginator:
    properties:
      limit:
        type: integer
      size:
        type: integer
      skip:
        type: integer
      total:
        type: integer
    type: object
  Project:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  ProjectShort:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  UpdateIssueRequest:
    properties:
      description:
        type: string
      dueDate:
        type: string
      priority:
        $ref: '#/definitions/IssuePriority'
      status:
        $ref: '#/definitions/IssueStatus'
      title:
        type: string
    type: object
  UpdateIssueResponse:
    properties:
      data:
        $ref: '#/definitions/Issue'
    type: object
  UpdateProjectRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  UpdateProjectResponse:
    properties:
      data:
        $ref: '#/definitions/Project'
    type: object
info:
  contact: {}
  description: API для управления проектами, задачами и комментариями.
  title: REST Demo API
  version: "1.0"
paths:
  /issues/{issueId}:
    delete:
      description: Delete a issue by its identifier.
      parameters:
      - description: Issue ID
        in: path
        minimum: 1
        name: issueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Issue deleted
      summary: Delete issue
      tags:
      - Issues
    get:
      description: Returns a issue by its identifier.
      parameters:
      - description: Issue ID
        in: path
        minimum: 1
        name: issueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetIssueResponse'
      summary: Get issue
      tags:
      - Issues
    patch:
      description: Update a issue.
      parameters:
      - description: Issue ID
        in: path
        name: issueId
        required: true
        type: integer
      - description: Issue data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateIssueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UpdateIssueResponse'
      summary: Update issue
      tags:
      - Issues
  /projects:
    get:
      description: Returns a projects list.
      parameters:
      - default: 0
        description: Number of projects to skip
        in: query
        minimum: 0
        name: skip
        type: integer
      - default: 10
        description: Maximum number of projects
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetProjectsListResponse'
      summary: Get projects list
      tags:
      - Projects
  /projects/{projectId}:
    delete:
      description: Delete a project by its identifier.
      parameters:
      - description: Project ID
        in: path
        minimum: 1
        name: projectId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Project deleted
      summary: Delete project
      tags:
      - Projects
    get:
      description: Returns a project by its identifier.
      parameters:
      - description: Project ID
        in: path
        minimum: 1
        name: projectId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetProjectResponse'
      summary: Get project
      tags:
      - Projects
    patch:
      description: Update a project.
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Project data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UpdateProjectResponse'
      summary: Update project
      tags:
      - Projects
    post:
      description: Create a project.
      parameters:
      - description: Project data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CreateProjectResponse'
      summary: Create project
      tags:
      - Projects
  /projects/{projectId}/issues:
    get:
      description: Returns a project issues list.
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - default: 0
        description: Number of projects to skip
        in: query
        minimum: 0
        name: skip
        type: integer
      - default: 10
        description: Maximum number of projects
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      - description: Issue status
        enum:
        - open
        - in_progress
        - done
        in: query
        name: status
        type: string
      - description: Issue priority
        enum:
        - low
        - medium
        - high
        in: query
        name: priority
        type: string
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetProjectIssuesListResponse'
      summary: Get project issues list
      tags:
      - Issues
    post:
      description: Create a project issue.
      parameters:
      - description: Project ID
        in: path
        minimum: 1
        name: projectId
        required: true
        type: integer
      - description: Issue data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateIssueRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CreateIssueResponse'
      summary: Create project issue
      tags:
      - Issues
produces:
- application/json
swagger: "2.0"